move dynamic to its own pkg (2)
