+++ pkg/api/validate/dynamic/dynamic.go
@@ -6,20 +6,24 @@ package dynamic
 import (
 	"context"
 	"fmt"
+	"net"
 	"net/http"
 	"sort"
 	"strings"
 	"time"
 
 	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2019-07-01/network"
+	mgmtfeatures "github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-07-01/features"
 	"github.com/Azure/go-autorest/autorest"
 	"github.com/Azure/go-autorest/autorest/azure"
+	"github.com/apparentlymart/go-cidr/cidr"
 	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/util/wait"
 
 	"github.com/Azure/ARO-RP/pkg/api"
 	"github.com/Azure/ARO-RP/pkg/env"
 	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/authorization"
+	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/features"
 	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/network"
 	"github.com/Azure/ARO-RP/pkg/util/permissions"
 	"github.com/Azure/ARO-RP/pkg/util/refreshable"
@@ -31,13 +35,16 @@ import (
 type Dynamic interface {
 	ValidateVnetPermissions(ctx context.Context) error
 	ValidateRouteTablesPermissions(ctx context.Context) error
-
+	ValidateCIDRRanges(ctx context.Context) error
+	ValidateVnetLocation(ctx context.Context) error
+	ValidateProviders(ctx context.Context) error
 	// etc
 	// does Quota code go in here too?
 }
 
 type dynamic struct {
 	log             *logrus.Entry
+	oc              *api.OpenShiftCluster
 	vnetr           *azure.Resource
 	masterSubnetID  string
 	workerSubnetIDs []string
@@ -46,10 +53,11 @@ type dynamic struct {
 	typ  string
 
 	permissions     authorization.PermissionsClient
+	providers       features.ProvidersClient
 	virtualNetworks virtualNetworksGetClient
 }
 
-func NewValidator(log *logrus.Entry, env env.Core, masterSubnetID string, workerSubnetIDs []string, subscriptionID string, authorizer refreshable.Authorizer, code string, typ string) (*dynamic, error) {
+func NewValidator(log *logrus.Entry, env env.Core, oc *api.OpenShiftCluster, masterSubnetID string, workerSubnetIDs []string, subscriptionID string, authorizer refreshable.Authorizer, code string, typ string) (*dynamic, error) {
 	vnetID, _, err := subnet.Split(masterSubnetID)
 	if err != nil {
 		return nil, err
@@ -62,6 +70,7 @@ func NewValidator(log *logrus.Entry, env env.Core, masterSubnetID string, worker
 
 	return &dynamic{
 		log:             log,
+		oc:              oc,
 		vnetr:           &vnetr,
 		masterSubnetID:  masterSubnetID,
 		workerSubnetIDs: workerSubnetIDs,
@@ -70,6 +79,7 @@ func NewValidator(log *logrus.Entry, env env.Core, masterSubnetID string, worker
 		typ:  typ,
 
 		permissions:     authorization.NewPermissionsClient(env.Environment(), subscriptionID, authorizer),
+		providers:       features.NewProvidersClient(env.Environment(), subscriptionID, authorizer),
 		virtualNetworks: newVirtualNetworksCache(network.NewVirtualNetworksClient(env.Environment(), subscriptionID, authorizer)),
 	}, nil
 }
@@ -196,6 +206,168 @@ func (dv *dynamic) validateActions(ctx context.Context, r *azure.Resource, actio
 
 		return true, nil
 	}, timeoutCtx.Done())
+}
+
+func (dv *dynamic) ValidateCIDRRanges(ctx context.Context) error {
+	dv.log.Print("ValidateCIDRRanges")
+
+	vnet, err := dv.virtualNetworks.Get(ctx, dv.vnetr.ResourceGroup, dv.vnetr.ResourceName, "")
+	if err != nil {
+		return err
+	}
+
+	var subnets []string
+	var CIDRArray []*net.IPNet
+
+	// unique names of subnets from all node pools
+	for i, subnet := range dv.oc.Properties.WorkerProfiles {
+		exists := false
+		for _, s := range subnets {
+			if strings.EqualFold(strings.ToLower(subnet.SubnetID), strings.ToLower(s)) {
+				exists = true
+				break
+			}
+		}
+
+		if !exists {
+			subnets = append(subnets, subnet.SubnetID)
+			path := fmt.Sprintf("properties.workerProfiles[%d].subnetId", i)
+			c, err := dv.validateSubnet(ctx, &vnet, path, subnet.SubnetID)
+			if err != nil {
+				return err
+			}
+
+			CIDRArray = append(CIDRArray, c)
+		}
+	}
+
+	masterCIDR, err := dv.validateSubnet(ctx, &vnet, "properties.MasterProfile.subnetId", dv.oc.Properties.MasterProfile.SubnetID)
+	if err != nil {
+		return err
+	}
+	_, podCIDR, err := net.ParseCIDR(dv.oc.Properties.NetworkProfile.PodCIDR)
+	if err != nil {
+		return err
+	}
+
+	_, serviceCIDR, err := net.ParseCIDR(dv.oc.Properties.NetworkProfile.ServiceCIDR)
+	if err != nil {
+		return err
+	}
+
+	CIDRArray = append(CIDRArray, masterCIDR, podCIDR, serviceCIDR)
+
+	err = cidr.VerifyNoOverlap(CIDRArray, &net.IPNet{IP: net.IPv4zero, Mask: net.IPMask(net.IPv4zero)})
+	if err != nil {
+		return api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, "", "The provided CIDRs must not overlap: '%s'.", err)
+	}
+
+	return nil
+}
+
+func (dv *dynamic) ValidateVnetLocation(ctx context.Context) error {
+	dv.log.Print("ValidateVnetLocation")
+
+	vnet, err := dv.virtualNetworks.Get(ctx, dv.vnetr.ResourceGroup, dv.vnetr.ResourceName, "")
+	if err != nil {
+		return err
+	}
+
+	if !strings.EqualFold(*vnet.Location, dv.oc.Location) {
+		return api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, "", "The vnet location '%s' must match the cluster location '%s'.", *vnet.Location, dv.oc.Location)
+	}
+
+	return nil
+}
+
+func (dv *dynamic) validateSubnet(ctx context.Context, vnet *mgmtnetwork.VirtualNetwork, path, subnetID string) (*net.IPNet, error) {
+	dv.log.Printf("validateSubnet (%s)", path)
+
+	s := findSubnet(vnet, subnetID)
+	if s == nil {
+		return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' could not be found.", subnetID)
+	}
+
+	if strings.EqualFold(dv.oc.Properties.MasterProfile.SubnetID, subnetID) {
+		if s.PrivateLinkServiceNetworkPolicies == nil ||
+			!strings.EqualFold(*s.PrivateLinkServiceNetworkPolicies, "Disabled") {
+			return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must have privateLinkServiceNetworkPolicies disabled.", subnetID)
+		}
+	}
+
+	var found bool
+	if s.ServiceEndpoints != nil {
+		for _, se := range *s.ServiceEndpoints {
+			if strings.EqualFold(*se.Service, "Microsoft.ContainerRegistry") &&
+				se.ProvisioningState == mgmtnetwork.Succeeded {
+				found = true
+				break
+			}
+		}
+	}
+	if !found {
+		return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must have Microsoft.ContainerRegistry serviceEndpoint.", subnetID)
+	}
+
+	if dv.oc.Properties.ProvisioningState == api.ProvisioningStateCreating {
+		if s.SubnetPropertiesFormat != nil &&
+			s.SubnetPropertiesFormat.NetworkSecurityGroup != nil {
+			return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must not have a network security group attached.", subnetID)
+		}
+
+	} else {
+		nsgID, err := subnet.NetworkSecurityGroupID(dv.oc, *s.ID)
+		if err != nil {
+			return nil, err
+		}
+
+		if s.SubnetPropertiesFormat == nil ||
+			s.SubnetPropertiesFormat.NetworkSecurityGroup == nil ||
+			!strings.EqualFold(*s.SubnetPropertiesFormat.NetworkSecurityGroup.ID, nsgID) {
+			return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must have network security group '%s' attached.", subnetID, nsgID)
+		}
+	}
+
+	_, net, err := net.ParseCIDR(*s.AddressPrefix)
+	if err != nil {
+		return nil, err
+	}
+
+	ones, _ := net.Mask.Size()
+	if ones > 27 {
+		return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must be /27 or larger.", subnetID)
+	}
+
+	return net, nil
+}
+
+func (dv *dynamic) ValidateProviders(ctx context.Context) error {
+	dv.log.Print("ValidateProviders")
+
+	providers, err := dv.providers.List(ctx, nil, "")
+	if err != nil {
+		return err
+	}
+
+	providerMap := make(map[string]mgmtfeatures.Provider, len(providers))
+
+	for _, provider := range providers {
+		providerMap[*provider.Namespace] = provider
+	}
+
+	for _, provider := range []string{
+		"Microsoft.Authorization",
+		"Microsoft.Compute",
+		"Microsoft.Network",
+		"Microsoft.Storage",
+	} {
+		if providerMap[provider].RegistrationState == nil ||
+			*providerMap[provider].RegistrationState != "Registered" {
+			return api.NewCloudError(http.StatusBadRequest, api.CloudErrorResourceProviderNotRegistered, "", "The resource provider '%s' is not registered.", provider)
+		}
+	}
+
+	return nil
 }
 
 func getRouteTableID(vnet *mgmtnetwork.VirtualNetwork, path, subnetID string) (string, error) {
