+++ pkg/api/validate/openshiftcluster_validatedynamic.go
@@ -5,14 +5,8 @@ package validate
 
 import (
 	"context"
-	"fmt"
-	"net"
 	"net/http"
-	"strings"
 
-	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2019-07-01/network"
-	mgmtfeatures "github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-07-01/features"
-	"github.com/apparentlymart/go-cidr/cidr"
 	"github.com/form3tech-oss/jwt-go"
 	"github.com/sirupsen/logrus"
 
@@ -20,9 +14,8 @@ import (
 	"github.com/Azure/ARO-RP/pkg/api/validate/dynamic"
 	"github.com/Azure/ARO-RP/pkg/env"
 	"github.com/Azure/ARO-RP/pkg/util/aad"
-	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/features"
+	"github.com/Azure/ARO-RP/pkg/util/azureclaim"
 	"github.com/Azure/ARO-RP/pkg/util/refreshable"
-	"github.com/Azure/ARO-RP/pkg/util/subnet"
 )
 
 // OpenShiftClusterDynamicValidator is the dynamic validator interface
@@ -39,16 +32,7 @@ func NewOpenShiftClusterDynamicValidator(log *logrus.Entry, env env.Core, oc *ap
 		oc:              oc,
 		subscriptionDoc: subscriptionDoc,
 		fpAuthorizer:    fpAuthorizer,
-		providers:       features.NewProvidersClient(env.Environment(), subscriptionDoc.ID, fpAuthorizer),
 	}
-}
-
-type azureClaim struct {
-	Roles []string `json:"roles,omitempty"`
-}
-
-func (*azureClaim) Valid() error {
-	return fmt.Errorf("unimplemented")
 }
 
 type openShiftClusterDynamicValidator struct {
@@ -58,7 +42,6 @@ type openShiftClusterDynamicValidator struct {
 	oc              *api.OpenShiftCluster
 	subscriptionDoc *api.SubscriptionDocument
 	fpAuthorizer    refreshable.Authorizer
-	providers       features.ProvidersClient
 }
 
 // Dynamic validates an OpenShift cluster
@@ -72,7 +55,7 @@ func (dv *openShiftClusterDynamicValidator) Dynamic(ctx context.Context) error {
 	}
 
 	// FP validation
-	fpDynamic, err := dynamic.NewValidator(dv.log, dv.env, mSubnetID, wSubnetIDs, dv.subscriptionDoc.ID, dv.fpAuthorizer, api.CloudErrorCodeInvalidResourceProviderPermissions, "resource provider")
+	fpDynamic, err := dynamic.NewValidator(dv.log, dv.env, dv.oc, mSubnetID, wSubnetIDs, dv.subscriptionDoc.ID, dv.fpAuthorizer, api.CloudErrorCodeInvalidResourceProviderPermissions, "resource provider")
 	if err != nil {
 		return err
 	}
@@ -101,7 +84,7 @@ func (dv *openShiftClusterDynamicValidator) Dynamic(ctx context.Context) error {
 
 	spAuthorizer := refreshable.NewAuthorizer(token)
 
-	spDynamic, err := dynamic.NewValidator(dv.log, dv.env, mSubnetID, wSubnetIDs, dv.subscriptionDoc.ID, spAuthorizer, api.CloudErrorCodeInvalidServicePrincipalPermissions, "provided service principal")
+	spDynamic, err := dynamic.NewValidator(dv.log, dv.env, dv.oc, mSubnetID, wSubnetIDs, dv.subscriptionDoc.ID, spAuthorizer, api.CloudErrorCodeInvalidServicePrincipalPermissions, "provided service principal")
 	if err != nil {
 		return err
 	}
@@ -117,176 +100,14 @@ func (dv *openShiftClusterDynamicValidator) Dynamic(ctx context.Context) error {
 	}
 
 	// Additional checks - use any dynamic because they both have the correct permissions
-	vnet, err := spDynamic.virtualNetworks.Get(ctx, spDynamic.vnetr.ResourceGroup, spDynamic.vnetr.ResourceName, "")
+	err = spDynamic.ValidateVnetLocation(ctx)
 	if err != nil {
 		return err
 	}
 
-	err = dv.validateCIDRRanges(ctx, &vnet)
+	err = spDynamic.ValidateProviders(ctx)
 	if err != nil {
 		return err
-	}
-
-	err = dv.validateVnetLocation(ctx, &vnet)
-	if err != nil {
-		return err
-	}
-
-	err = dv.validateProviders(ctx)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (dv *openShiftClusterDynamicValidator) validateVnetLocation(ctx context.Context, vnet *mgmtnetwork.VirtualNetwork) error {
-	dv.log.Print("validateVnetLocation")
-
-	if !strings.EqualFold(*vnet.Location, dv.oc.Location) {
-		return api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, "", "The vnet location '%s' must match the cluster location '%s'.", *vnet.Location, dv.oc.Location)
-	}
-
-	return nil
-}
-
-func (dv *openShiftClusterDynamicValidator) validateCIDRRanges(ctx context.Context, vnet *mgmtnetwork.VirtualNetwork) error {
-	dv.log.Print("validateCIDRRanges")
-
-	var subnets []string
-	var CIDRArray []*net.IPNet
-
-	// unique names of subnets from all node pools
-	for i, subnet := range dv.oc.Properties.WorkerProfiles {
-		exists := false
-		for _, s := range subnets {
-			if strings.EqualFold(strings.ToLower(subnet.SubnetID), strings.ToLower(s)) {
-				exists = true
-				break
-			}
-		}
-
-		if !exists {
-			subnets = append(subnets, subnet.SubnetID)
-			path := fmt.Sprintf("properties.workerProfiles[%d].subnetId", i)
-			c, err := dv.validateSubnet(ctx, vnet, path, subnet.SubnetID)
-			if err != nil {
-				return err
-			}
-
-			CIDRArray = append(CIDRArray, c)
-		}
-	}
-
-	masterCIDR, err := dv.validateSubnet(ctx, vnet, "properties.MasterProfile.subnetId", dv.oc.Properties.MasterProfile.SubnetID)
-	if err != nil {
-		return err
-	}
-	_, podCIDR, err := net.ParseCIDR(dv.oc.Properties.NetworkProfile.PodCIDR)
-	if err != nil {
-		return err
-	}
-
-	_, serviceCIDR, err := net.ParseCIDR(dv.oc.Properties.NetworkProfile.ServiceCIDR)
-	if err != nil {
-		return err
-	}
-
-	CIDRArray = append(CIDRArray, masterCIDR, podCIDR, serviceCIDR)
-
-	err = cidr.VerifyNoOverlap(CIDRArray, &net.IPNet{IP: net.IPv4zero, Mask: net.IPMask(net.IPv4zero)})
-	if err != nil {
-		return api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, "", "The provided CIDRs must not overlap: '%s'.", err)
-	}
-
-	return nil
-}
-
-func (dv *openShiftClusterDynamicValidator) validateSubnet(ctx context.Context, vnet *mgmtnetwork.VirtualNetwork, path, subnetID string) (*net.IPNet, error) {
-	dv.log.Printf("validateSubnet (%s)", path)
-
-	s := findSubnet(vnet, subnetID)
-	if s == nil {
-		return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' could not be found.", subnetID)
-	}
-
-	if strings.EqualFold(dv.oc.Properties.MasterProfile.SubnetID, subnetID) {
-		if s.PrivateLinkServiceNetworkPolicies == nil ||
-			!strings.EqualFold(*s.PrivateLinkServiceNetworkPolicies, "Disabled") {
-			return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must have privateLinkServiceNetworkPolicies disabled.", subnetID)
-		}
-	}
-
-	var found bool
-	if s.ServiceEndpoints != nil {
-		for _, se := range *s.ServiceEndpoints {
-			if strings.EqualFold(*se.Service, "Microsoft.ContainerRegistry") &&
-				se.ProvisioningState == mgmtnetwork.Succeeded {
-				found = true
-				break
-			}
-		}
-	}
-	if !found {
-		return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must have Microsoft.ContainerRegistry serviceEndpoint.", subnetID)
-	}
-
-	if dv.oc.Properties.ProvisioningState == api.ProvisioningStateCreating {
-		if s.SubnetPropertiesFormat != nil &&
-			s.SubnetPropertiesFormat.NetworkSecurityGroup != nil {
-			return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must not have a network security group attached.", subnetID)
-		}
-
-	} else {
-		nsgID, err := subnet.NetworkSecurityGroupID(dv.oc, *s.ID)
-		if err != nil {
-			return nil, err
-		}
-
-		if s.SubnetPropertiesFormat == nil ||
-			s.SubnetPropertiesFormat.NetworkSecurityGroup == nil ||
-			!strings.EqualFold(*s.SubnetPropertiesFormat.NetworkSecurityGroup.ID, nsgID) {
-			return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must have network security group '%s' attached.", subnetID, nsgID)
-		}
-	}
-
-	_, net, err := net.ParseCIDR(*s.AddressPrefix)
-	if err != nil {
-		return nil, err
-	}
-
-	ones, _ := net.Mask.Size()
-	if ones > 27 {
-		return nil, api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidLinkedVNet, path, "The provided subnet '%s' is invalid: must be /27 or larger.", subnetID)
-	}
-
-	return net, nil
-}
-
-func (dv *openShiftClusterDynamicValidator) validateProviders(ctx context.Context) error {
-	dv.log.Print("validateProviders")
-
-	providers, err := dv.providers.List(ctx, nil, "")
-	if err != nil {
-		return err
-	}
-
-	providerMap := make(map[string]mgmtfeatures.Provider, len(providers))
-
-	for _, provider := range providers {
-		providerMap[*provider.Namespace] = provider
-	}
-
-	for _, provider := range []string{
-		"Microsoft.Authorization",
-		"Microsoft.Compute",
-		"Microsoft.Network",
-		"Microsoft.Storage",
-	} {
-		if providerMap[provider].RegistrationState == nil ||
-			*providerMap[provider].RegistrationState != "Registered" {
-			return api.NewCloudError(http.StatusBadRequest, api.CloudErrorResourceProviderNotRegistered, "", "The resource provider '%s' is not registered.", provider)
-		}
 	}
 
 	return nil
@@ -304,7 +125,7 @@ func validateServicePrincipalProfile(ctx context.Context, log *logrus.Entry, env
 	}
 
 	p := &jwt.Parser{}
-	c := &azureClaim{}
+	c := &azureclaim.AzureClaim{}
 	_, _, err = p.ParseUnverified(token.OAuthToken(), c)
 	if err != nil {
 		return err
@@ -313,18 +134,6 @@ func validateServicePrincipalProfile(ctx context.Context, log *logrus.Entry, env
 	for _, role := range c.Roles {
 		if role == "Application.ReadWrite.OwnedBy" {
 			return api.NewCloudError(http.StatusBadRequest, api.CloudErrorCodeInvalidServicePrincipalCredentials, "properties.servicePrincipalProfile", "The provided service principal must not have the Application.ReadWrite.OwnedBy permission.")
-		}
-	}
-
-	return nil
-}
-
-func findSubnet(vnet *mgmtnetwork.VirtualNetwork, subnetID string) *mgmtnetwork.Subnet {
-	if vnet.Subnets != nil {
-		for _, s := range *vnet.Subnets {
-			if strings.EqualFold(*s.ID, subnetID) {
-				return &s
-			}
 		}
 	}
